Move AAC silent frame stuff into separate AAC helper.
