+++ src/remux/mp4-remuxer.js
@@ -3,7 +3,7 @@
 */
 
 
-import ADTS from '../demux/adts';
+import AAC from '../helper/aac';
 import Event from '../events';
 import {logger} from '../utils/logger';
 import MP4 from '../remux/mp4-generator';
@@ -337,7 +337,7 @@ class MP4Remuxer {
         logger.log(`Injecting ${missing} packets of missing audio.`);
         for (var j = 0; j < missing; j++) {
           var newStamp = Math.round(sample0.pts + (j + 1) * 1024 * pes2mp4ScaleFactor),
-              fillFrame = ADTS.getSilentFrame(track.channelCount),
+              fillFrame = AAC.getSilentFrame(track.channelCount),
               newAacSample;
           if (!fillFrame) {
             logger.log('Unable to get silent frame for given audio codec; duplicating last frame instead.');
@@ -473,7 +473,7 @@ class MP4Remuxer {
         nbSamples = Math.ceil((endDTS - startDTS) / frameDuration),
 
         // silent frame
-        silentFrame = ADTS.getSilentFrame(track.channelCount);
+        silentFrame = AAC.getSilentFrame(track.channelCount);
 
     // Can't remux if we can't generate a silent frame...
     if (!silentFrame) {
